Aggressive inlining of functions forced me to pull in too many headers and slowed down compilation to a crawl. From now on, the header that needs to be inclued in profiled files only contains the definitions of vital macros and functions. This extremely lightly slightly slows down the execution, but makes the compilation many times faster.
The core functionality (e.g., result saving and output, thread id assignment, etc.) was moved to a separate header that should only be included in files that actually need the said functionality.

Updated the macros to use an IYFT prefix instead of an IYF prefix for consistency. Improved them as well - no more warnings are reported.

Initial support for drawing results in imgui.

Tons of other fixes and improvements.